Consolidated constants for better auditing when fixing other tickets
