+++ packages/cydran/src/Constants.ts
@@ -4,6 +4,62 @@ enum Ids {
 	STAGE_INTERNALS = "cydranStageInternals"
 }
 
+enum Attrs {
+	ID = "id",
+	NAME = "name",
+	COMPONENT = "component",
+	CONTEXT = "context",
+	VALUE = "value",
+	LOCK = "lock",
+	TYPE = "type",
+	TEST = "test"
+}
+
+enum CydranMode {
+	STRICT = "strict",
+	LAZY = "lazy"
+}
+
+enum DigestionActions {
+	REQUEST_DIGESTION_SOURCES = "REQUEST_DIGESTION_SOURCES",
+	REQUEST_DIGESTION_CANDIDATES = "REQUEST_DIGESTION_CANDIDATES"
+}
+
+enum TagNames {
+	SCRIPT = "script",
+	TEMPLATE = "template",
+	CYDRAN_REGION = "c-region",
+	CYDRAN_COMPONENT_STYLES = "c-component-styles"
+}
+
+const JSType = {
+	STR: "string",
+	BOOL: "boolean",
+	BIGINT: "bigint",
+	NUM: "number",
+	SYM: "symbol",
+	FN: "function",
+	OBJ: "object",
+	UND: "undefined"
+} as const;
+
+enum Events {
+	AFTER_CHILD_ADDED = "AFTER_CHILD_ADDED",
+	AFTER_CHILD_CHANGED = "AFTER_CHILD_CHANGED",
+	AFTER_CHILD_REMOVED = "AFTER_CHILD_REMOVED",
+	AFTER_PARENT_ADDED = "AFTER_PARENT_ADDED",
+	AFTER_PARENT_CHANGED = "AFTER_PARENT_CHANGED",
+	AFTER_PARENT_REMOVED = "AFTER_PARENT_REMOVED",
+	BEFORE_CHILD_ADDED = "BEFORE_CHILD_ADDED",
+	BEFORE_CHILD_CHANGED = "BEFORE_CHILD_CHANGED",
+	BEFORE_CHILD_REMOVED = "BEFORE_CHILD_REMOVED",
+	BEFORE_PARENT_ADDED = "BEFORE_PARENT_ADDED",
+	BEFORE_PARENT_CHANGED = "BEFORE_PARENT_CHANGED",
+	BEFORE_PARENT_REMOVED = "BEFORE_PARENT_REMOVED",
+	COMPONENT_NESTING_CHANGED = "COMPONENT_NESTING_CHANGED",
+	CYDRAN_PREAPP_DISPOSAL = "CYDRAN_PREAPP_DISPOSAL"
+}
+
 enum PropertyKeys {
 	CYDRAN_CLONE_MAX_EVALUATIONS = "cydran.clone.maxEvaluations",
 	CYDRAN_DIGEST_MAX_EVALUATIONS = "cydran.digest.maxEvaluations",
@@ -23,6 +79,7 @@ enum PropertyKeys {
 	CYDRAN_STYLES_ENABLED = "cydran.styles.enabled"
 }
 
+const DEFAULT_ID_KEY: string = "id";
 const ANONYMOUS_REGION_PREFIX: string = "%%%Region_" as const;
 const ATTRIBUTE_DELIMITER: string = "-" as const;
 const BEHAVIOR_FORM_RESET: string = "cydran:behavior:form:reset" as const;
@@ -49,26 +106,33 @@ const VALID_KEY: RegExp = new RegExp(/^[a-zA-Z\$\_][a-zA-Z\d\$\_\*]*$/);
 export {
 	ANONYMOUS_REGION_PREFIX,
 	ATTRIBUTE_DELIMITER,
+	Attrs,
 	BEHAVIOR_FORM_RESET,
 	CHANGE_KEY,
 	CYDRAN_DISPOSE_FN_NAME,
 	CYDRAN_KEY,
 	CYDRAN_PUBLIC_CHANNEL,
 	CYDRAN_SCRIPT_PREFIX,
+	CydranMode,
 	DEFAULT_CLONE_DEPTH,
 	DEFAULT_CONTEXT_KEY,
 	DEFAULT_EQUALS_DEPTH,
+	DEFAULT_ID_KEY,
 	DEFAULT_LOG_STRATEGY,
 	DEFAULT_PREFIX,
 	DOM_KEY,
+	DigestionActions,
+	Events,
 	INPUT_KEY,
 	INTERNAL_CHANNEL_NAME,
 	Ids,
+	JSType,
 	LITERAL_PATH_REGEX,
 	LOCAL_ID_REGEX,
 	PropertyKeys,
 	RELATIVE_PATH_REGEX,
 	RESET_KEY,
+	TagNames,
 	VALID_ID,
 	VALID_KEY
 };
