+++ packages/cydran/src/behavior/BehaviorInternalsImpl.ts
@@ -8,7 +8,7 @@ import Machine from "machine/Machine";
 import MachineState from "machine/MachineState";
 import Behavior from "behavior/Behavior";
 import stateMachineBuilder from "machine/StateMachineBuilder";
-import { VALID_ID, DOM_KEY, INTERNAL_CHANNEL_NAME } from "Constants";
+import { VALID_ID, DOM_KEY, INTERNAL_CHANNEL_NAME, DigestionActions } from "Constants";
 import { requireNotNull, isDefined, requireValid, elementAsString, hasContents, defaulted } from 'util/Utils';
 import SimpleMap from "interface/SimpleMap";
 import Attributes from "component/Attributes";
@@ -21,7 +21,6 @@ import BehaviorAttributeConverters from "behavior/BehaviorAttributeConverters";
 import AttributeParser from 'validator/AttributeParser';
 import AttributeParserImpl from "validator/AttributeParserImpl";
 import { asIdentity } from "util/AsFunctions";
-import DigestionActions from "const/DigestionActions";
 import { BehaviorError, ContextUnavailableError } from "error/Errors";
 import InternalBehaviorFlags from "behavior/InternalBehaviorFlags";
 import OnContinuation from "continuation/OnContinuation";
