--- packages/cydran/src/const/AttrsFields.ts
@@ -1,12 +1,0 @@
-enum Attrs {
-	ID = "id",
-	NAME = "name",
-	COMPONENT = "component",
-	CONTEXT = "context",
-	VALUE = "value",
-	LOCK = "lock",
-	TYPE = "type",
-	TEST = "test"
-}
-
-export default Attrs;
