--- packages/cydran/src/const/CydranMode.ts
@@ -1,6 +1,0 @@
-enum CydranMode {
-	STRICT = "strict",
-	LAZY = "lazy"
-}
-
-export default CydranMode;
