--- packages/cydran/src/const/DefaultIdKey.ts
@@ -1,3 +1,0 @@
-const DEFAULT_ID_KEY: string = "id";
-
-export default DEFAULT_ID_KEY;
