--- packages/cydran/src/const/DigestionActions.ts
@@ -1,6 +1,0 @@
-enum DigestionActions {
-	REQUEST_DIGESTION_SOURCES = "REQUEST_DIGESTION_SOURCES",
-	REQUEST_DIGESTION_CANDIDATES = "REQUEST_DIGESTION_CANDIDATES"
-}
-
-export default DigestionActions;
