--- packages/cydran/src/const/EventsFields.ts
@@ -1,18 +1,0 @@
-enum Events {
-	AFTER_CHILD_ADDED = "AFTER_CHILD_ADDED",
-	AFTER_CHILD_CHANGED = "AFTER_CHILD_CHANGED",
-	AFTER_CHILD_REMOVED = "AFTER_CHILD_REMOVED",
-	AFTER_PARENT_ADDED = "AFTER_PARENT_ADDED",
-	AFTER_PARENT_CHANGED = "AFTER_PARENT_CHANGED",
-	AFTER_PARENT_REMOVED = "AFTER_PARENT_REMOVED",
-	BEFORE_CHILD_ADDED = "BEFORE_CHILD_ADDED",
-	BEFORE_CHILD_CHANGED = "BEFORE_CHILD_CHANGED",
-	BEFORE_CHILD_REMOVED = "BEFORE_CHILD_REMOVED",
-	BEFORE_PARENT_ADDED = "BEFORE_PARENT_ADDED",
-	BEFORE_PARENT_CHANGED = "BEFORE_PARENT_CHANGED",
-	BEFORE_PARENT_REMOVED = "BEFORE_PARENT_REMOVED",
-	COMPONENT_NESTING_CHANGED = "COMPONENT_NESTING_CHANGED",
-	CYDRAN_PREAPP_DISPOSAL = "CYDRAN_PREAPP_DISPOSAL"
-}
-
-export default Events;
