--- packages/cydran/src/const/JSType.ts
@@ -1,12 +1,0 @@
-const JSType = {
-	STR: "string",
-	BOOL: "boolean",
-	BIGINT: "bigint",
-	NUM: "number",
-	SYM: "symbol",
-	FN: "function",
-	OBJ: "object",
-	UND: "undefined"
-} as const;
-
-export default JSType;
