--- packages/cydran/src/const/TagNames.ts
@@ -1,8 +1,0 @@
-enum TagNames {
-	SCRIPT = "script",
-	TEMPLATE = "template",
-	CYDRAN_REGION = "c-region",
-	CYDRAN_COMPONENT_STYLES = "c-component-styles"
-}
-
-export default TagNames;
