+++ packages/cydran/src/index.ts
@@ -1,4 +1,4 @@
-import { CYDRAN_KEY, PropertyKeys } from "Constants";
+import { JSType, Events, CYDRAN_KEY, PropertyKeys } from "Constants";
 import ArgumentsResolvers from 'argument/ArgumentsResolvers';
 import AbstractBehavior from "behavior/AbstractBehavior";
 import AbstractValueBehavior from "behavior/AbstractValueBehavior";
@@ -12,8 +12,6 @@ import ElementOperations from "component/ElementOperations";
 import MetadataContinuation from "component/MetadataContinuation";
 import Renderer from "component/Renderer";
 import argumentsBuilder from "function/argumentsBuilder";
-import Events from "const/EventsFields";
-import JSType from "const/JSType";
 import { Context, Stage } from "context/Context";
 import GlobalContextImpl from "context/GlobalContextImpl";
 import ForChannelContinuation from "continuation/ForChannelContinuation";
