+++ packages/cydran/src/scratch.ts
@@ -1,6 +1,5 @@
 // This is for things that need to be reintegrated after the refactor
-import Events from 'const/EventsFields';
-import { CYDRAN_PUBLIC_CHANNEL } from 'Constants';
+import { Events, CYDRAN_PUBLIC_CHANNEL } from 'Constants';
 
 // this.addDisposer((stage: Stage) => {
 // 	stage.sendGlobally(CYDRAN_PUBLIC_CHANNEL, Events.CYDRAN_PREAPP_DISPOSAL);
