+++ packages/cydran/src/stage/StageInternalsImpl.ts
@@ -8,9 +8,8 @@ import ComponentIdPair from 'component/CompnentIdPair';
 import MachineState from 'machine/MachineState';
 import Logger from 'log/Logger';
 import SimpleMap from 'interface/SimpleMap';
-import { PropertyKeys, Ids } from 'Constants';
+import { CydranMode, PropertyKeys, Ids } from 'Constants';
 import ContextTransitions from 'component/ContextTransitions';
-import CydranMode from 'const/CydranMode';
 import ContextStates from 'component/ContextStates';
 import LoggerFactory from 'log/LoggerFactory';
 import DomUtils from 'dom/DomUtils';
